// frontend/src/models/groupsModelType.ts
export interface groupsModelType extends Logic {
// ...
    selectors: {
        groupTypes: (state: any, props?: any) => Array<GroupType>
        groupTypesLoading: (state: any, props?: any) => boolean
        currentTeamId: (state: any, props?: any) => number | null
// ...
    }
// ...
}